Update .dir-locals.el
When attempting to use a CIDER REPL with a newly created project from this template, the classpath was not set correctly. Rather than including the `/src` and `/resources` paths, only the project root was on the classpath. I realized that the inclusion of this `:build` alias via `.dir-locals.el` was causing the `:paths` to be replaced by `.` (as specified in the `:build` alias). This was surprising for me, and I think it may surprise others as well, as I wasn't able to `(require '[...])` namespaces in the usual way after the REPL started.

Signed-off-by: Robert Medeiros <robert@crimeminister.org>